Replace script process with the instance one - don't fork
Previous solution tracking the forked process didn't do well, upstart
didn't do a management right. Even with the `expect fork`, it was still
expected from the shell (the parent process) to exit, and the upstart
never gets notification that the parent process exited.

This is now solved with replacing the process image calling exec(1)
inside script and removing the expect fork stanza. This replaces the
shell without creating a new process, making upstart tracing trivial.

Note, we still need to execute this inside `script`, as we need to
expand `INSTANCE` env. variable, which is possible only inside script
block.