Convert storage iterators to work with hash_t internally
They are now only rendered to strings where required by the request
protocols (see dhtproto).

A state enum is added to StorageEngineStepIterator, because we now
not only need to know if the iteration has started, but also if it
has finished. (Before, the finished state was represented by an
empty hash string.)

Fixes #119.