Import HashRange directly from swarm in RedistDhtClient
`HashRange` is not a public import in `dhtproto.client.legacy.DhtConst`,
so it is only accessible thanks to a compiler bug which makes selective
imports implicitly public.

This bug has been fixed in more recent D2 compiler versions, so we need
to import from swarm to guarantee access to the `HashRange` symbol.