Add a StompingPreventionStats version constraint
The `stomping_prevention_counter` global is only accessible when using
`dmd-transitional`.  If we want to build with vanilla DMD or any other
compiler, we cannot use this symbol, so this patch replaces the original
`D_Version2` version constraint with one specific to stomping prevention
stats.

By default this will be enabled in `Build.mak`, but it can be overridden
by using the `DISABLE_STOMPING_STATS` environment variable.  This should
ensure that current behaviour persists by default, but build for vanilla
DMD (or other compilers) can be added which disable the feature.